fix: Solved the onPageLoad() -> turnButtons() green!
Just as I suspected the list-items were not being populated at pageLoad.

Wish I had believed in myself and tested it out earlier.

For now, I have just added a delay to allow for the list-items to be
populated before I fetch them in turnButtons(). This I feel is very
hacky!

Will look for an event which is triggered when list-items are loaded and
add an event listener to that.